ai-form-filler-backend/requirements.txt
fastapi
uvicorn
groq
PyPDF2
aiosqlite
python-multipart
pydantic
python-dotenv
